Create DB with tables “Talents”, “ TalentInfo” / test connection with DB#10

+++ src/main/java/com/softserve/skillscope/talent/TalentRepository.java
@@ -2,6 +2,9 @@ package com.softserve.skillscope.talent;
 
 import com.softserve.skillscope.talent.model.entity.Talent;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Modifying;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.transaction.annotation.Transactional;
 
-public interface TalentRepository extends JpaRepository<Talent, Integer> {
+public interface TalentRepository extends JpaRepository<Talent, Long> {
 }
